refactor api tests
